Move toHHMMSS to Utils.js
