+++ map/src/util/Utils.js
@@ -133,6 +133,24 @@ function hexToArgb(hex) {
     return `rgb(${red} ${green} ${blue}${alphaString})`;
 }
 
+export const toHHMMSS = function (time) {
+    var sec_num = time / 1000;
+    var hours = Math.floor(sec_num / 3600);
+    var minutes = Math.floor((sec_num - hours * 3600) / 60);
+    var seconds = sec_num - hours * 3600 - minutes * 60;
+
+    if (hours < 10) {
+        hours = '0' + hours;
+    }
+    if (minutes < 10) {
+        minutes = '0' + minutes;
+    }
+    if (seconds < 10) {
+        seconds = '0' + Math.round(seconds);
+    }
+    return hours + ':' + minutes + ':' + seconds;
+};
+
 /*
     Prepare string with NaN/Infinity before JSON.parse()
